# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de API, backend",
        "version": "0.1"
    },
    "paths": {
...
        "/ICECREAM/public/api/get-categorias": {
            "get": {
                "tags": [
                    "Categorias productos"
                ],
                "description": "Obtiene todas las categorias productos",
                "operationId": "d688e7c9969288a148065ad37fc4ffee",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "nombre tienda"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "paletas de hielo sin leche"
                                            },
                                            "fecha_registro": {
                                                "type": "string",
                                                "example": "2024-02-29 09:27:54"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": [
                                            {
                                                "id": 1,
                                                "nombre": "nombre paletas/hielo",
                                                "descripcion": "paletas de hielo sin leche",
                                                "fecha_registro": "2024-02-29 09:27:54"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario fuera de sesion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ICECREAM/public/api/creaCategoria": {
            "post": {
                "tags": [
                    "Categorias productos"
                ],
                "description": "Guarda tiendas",
                "operationId": "a9f31dd87918f7313aae7355c531167f",
                "requestBody": {
                    "description": "El cuerpo del body debe de tener la siguiente estructura",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "nombre": {
                                            "type": "string",
                                            "example": "nombre categoria"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "descripcion..."
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "examples": {
                                "result": {
                                    "summary": "",
                                    "value": {
                                        "nombre": "nombre tienda",
                                        "descripcion": "descripcion.."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": "..."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "En proceso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "success": false,
                                            "message": "erro",
                                            "data": "Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ICECREAM/public/api/editaCategoria": {
            "put": {
                "tags": [
                    "Categorias productos"
                ],
                "description": "Guarda tiendas",
                "operationId": "f3ec1a71e2c7045687bcb5b5dfd0ecf0",
                "requestBody": {
                    "description": "El cuerpo del body debe de tener la siguiente estructura",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "nombre": {
                                            "type": "string",
                                            "example": "nombre tienda"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "descripcion..."
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "examples": {
                                "result": {
                                    "summary": "",
                                    "value": {
                                        "nombre": "nombre tienda",
                                        "descripcion": "calle tal, numero tal",
                                        "id": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": "..."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "En proceso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "success": false,
                                            "message": "erro",
                                            "data": "Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}